chore: Remove unused code and update dependencies
