+++ OpenMusic-Customer_Rasya/producer.js
@@ -1,0 +1,22 @@
+const amqp = require('amqplib');
+
+const init = async () => {
+  const connection = await amqp.connect('amqp://admin:password@localhost:5672');
+  const channel = await connection.createChannel();
+
+  const queue = 'dicoding';
+  const message = 'Selamat belajar message broker!';
+
+  await channel.assertQueue(queue, {
+    durable: true,
+  });
+
+  await channel.sendToQueue(queue, Buffer.from(message));
+  console.log('Pesan berhasil terkirim!');
+
+  setTimeout(() => {
+    connection.close();
+  }, 1000);
+};
+
+init();
